+++ Code/Framework/AzFramework/Tests/terrain_mock_files.cmake
@@ -7,6 +7,5 @@
 #
 
 set(FILES
-    Mocks/Terrain/MockTerrainLayerSpawner.h
-    Mocks/Terrain/MockTerrainAreaSurfaceRequestBus.h
+    Mocks/Terrain/MockTerrainDataRequestBus.h
 )
