+++ Gems/Terrain/Code/Tests/TerrainSystemTest.cpp
@@ -15,7 +15,7 @@
 #include <Components/TerrainLayerSpawnerComponent.h>
 
 #include <GradientSignal/Ebuses/MockGradientRequestBus.h>
-#include <Terrain/MockTerrain.h>
+#include <Tests/Mocks/Terrain/MockTerrainDataRequestBus.h>
 #include <Terrain/MockTerrainAreaSurfaceRequestBus.h>
 #include <MockAxisAlignedBoxShapeComponent.h>
